backend/requirements.txt
fastapi[standard]==0.115.12
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
supabase==2.28.0
google-genai==1.56.0
deepgram-sdk==6.0.1
web3==7.14.1
weasyprint==63.1
jinja2==3.1.5
firebase-admin==6.5.0
python-telegram-bot==21.10
google-auth>=2.45.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.164.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
nltk==3.9.1
python-dotenv==1.0.1
phonenumbers==8.13.55
httpx==0.28.1
websockets==14.2
requests==2.32.3
bandit==1.8.3
pytest==8.3.5
pytest-asyncio==0.25.3
